pyproject.toml
[project]
name = "ande-reproduction"
version = "0.1.0"
description = "Reproduction of ANDE: Detect the Anonymity Web Traffic With Comprehensive Model (IEEE TNSM 2024)"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    # Deep learning
    "torch>=2.3",
    "torchvision>=0.18",
    # Numerics / data
    "numpy>=1.26",
    "pandas>=2.2",
    "pyarrow>=15.0",
    # PCAP / traffic parsing
    "scapy>=2.5",
    "dpkt>=1.9",
    # ML baselines
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",
    # Utilities
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "rich>=13.7",
    "tensorboard>=2.16",
]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
    "ipykernel>=6.29",
    "jupyter>=1.0",
]

# PyTorch wheels live on a separate index. CUDA 12.4 build for both
# Windows (local RTX 4060) and Linux (cloud GPU). Driver >= 550 required.
[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
torchvision = [{ index = "pytorch-cu124" }]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ande"]

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
